requirements.txt
playwright>=1.40,<2.0
playwright-stealth>=1.0
requests>=2.31,<3.0
pyyaml>=6.0,<7.0
pytest>=7.0,<9.0
